refactor(server): extract deployment-manager free functions into siblings
Everything above class DeploymentManager was free functions that never
touch the class's state, so reaching the class meant scrolling past 860
lines of host probing and token minting. deployment-manager.ts is the
single most expensive file to read in this repo by opens x length.

  host-capabilities.ts    systemd/nix probes, probe caches, worker signal
  conversation-locks.ts   reserved-lock cap, counters, lock acquisition
  worker-invocation.ts    bun/node resolution, shell quoting, nix attrs
  deployment-identity.ts  token minting, deployment naming, secret detection

The class itself is untouched and unsplit; its cohesion is why multi-replica
dispatch works. deployment-manager.ts re-exports every symbol it exported
before, so no importer changes.

Behaviour-preserving detail worth noting: the class previously demoted
systemd for the session by assigning cachedSystemdRun = null directly. That
state now lives in host-capabilities.ts, so it calls disableSystemdRunForSession().
The null/undefined distinction is load-bearing — null means probed and
unavailable, undefined means unprobed and re-probes — and is preserved
exactly, including in __resetCapabilityProbesForTests.

2,951 -> 2,300 lines.

(cherry picked from commit ec5499cfe516bbaa203f5e9f91c8b9ee2216b1af)
